Move forwardTo to metadata builder to share with edge builder
