+++ ngrok-java/src/main/java/com/ngrok/MetadataBuilder.java
@@ -9,6 +9,7 @@ import java.util.Optional;
  */
 public abstract class MetadataBuilder<T extends MetadataBuilder> {
     private Optional<String> metadata = Optional.empty();
+    private Optional<String> forwardsTo = Optional.empty();
 
     /**
      * Sets the metadata for this builder.
@@ -22,11 +23,32 @@ public abstract class MetadataBuilder<T extends MetadataBuilder> {
     }
 
     /**
+     * Sets the forwarding address for this builder.
+     *
+     * @param forwardsTo the forwarding address for the builder
+     * @return An instance the builder represented by type T
+     */
+    public T forwardsTo(String forwardsTo) {
+        this.forwardsTo = Optional.of(forwardsTo);
+        return (T) this;
+    }
+
+    /**
      * Returns the metadata for this builder.
      *
      * @return the metadata
      */
     public Optional<String> getMetadata() {
         return metadata;
+    }
+
+
+    /**
+     * Returns the forwarding address for this builder.
+     *
+     * @return the currently set forwarding address
+     */
+    public Optional<String> getForwardsTo() {
+        return forwardsTo;
     }
 }
